Path_In_Zigzag_Labelled_Binary_Tree

+++ src/leetcode/Path_In_Zigzag_Labelled_Binary_Tree.java
@@ -9,7 +9,7 @@ public class Path_In_Zigzag_Labelled_Binary_Tree {
 
     public List<Integer> pathInZigZagTree(int label) {
         int lev=1;
-        while (Math.pow(2,(double)lev)-1<label){
+        while (Math.pow(2,lev)-1<label){
             lev++;
         }
         List<Integer> res=new ArrayList<>();
@@ -17,8 +17,8 @@ public class Path_In_Zigzag_Labelled_Binary_Tree {
        while (lev>1){
            lev--;
            int font= res.get(res.size()-1);
-           int cha=(int)(font-Math.pow(2,(double)lev))/2;
-           int next=(int) Math.pow(2,(double)lev)-1-cha;
+           int cha=(int)(font-Math.pow(2,lev))/2;
+           int next=(int) Math.pow(2,lev)-1-cha;
            res.add(next);
        }
         Collections.reverse(res);
